+++ resources/aws_ec2_capacity_reservation_fleet.rb
@@ -1,7 +1,6 @@
 # Import API specifics
 use "awscc_base"
 
-unified_mode true
 resource_name :aws_ec2_capacity_reservation_fleet
 provides :aws_ec2_capacity_reservation_fleet, target_mode: true, platform: "aws"